More swift mailer progress
Messages are actually being sent and the process doesn't just die if a
message is sent with errors…the next go round any messages that haven't
been attempted will be tried.  Still needs work, but a status column is
present on the response page and the logs are available in a hover.
Getting there :)