More Swiftmailer integration
Matched up many of the new fields that swift mailer offer.  I removed
the title field, added a sender friendly name field.  Made it so that
you can put the link anywhere you want just like the @fname and @lname
variables.

Added new message icon statuses.  When you hover over them you see the
logs of that message.

There is still a problem with message delay and you have to hit refresh
to keep the email script running.  Need to figure out how to get it to
run automatically.

There is still a problem with some error messages that break the mail
script.

Need to figure out how to parse the email logs for errors and not
provide a false sense of success just because the message was sent.