Module Upload/Upgrade Improvements to satisfy Issue #69
I've built in the validation so that you cannot install a module with
the same name or path unless you specify it as an upgrade.  The upgrade
should be tested however.  I've also eliminated the need for the
module_cleanup.php file.  These changes will require changes in the
install.php file that is included with new modules.  However, it is now
much easier to write the install.php file.